+++ package.json
@@ -3,9 +3,12 @@
   "version": "0.1.1",
   "private": true,
   "dependencies": {
+    "core-js": "^3.6.2",
+    "mathjs": "^6.5.0",
+    "node-sass": "^4.13.0",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
-    "react-scripts": "3.1.1"
+    "react-scripts": "^3.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
