+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import "./App.css";
-import Calculator from './components/Calculator';
+import Calculator from './components/Calculator/Calculator';
 // STEP 4 - import the button and display components
 // Don't forget to import any extra css/scss files you build into the correct component
