+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import logo from "../../Img/Lambda_Logo_white.png";
+import logo from "../Img/Lambda_Logo_white.png";
 
 const Header = () => {
   return (
