+++ src/data.js
@@ -31,4 +31,4 @@ const operators = [
 
 const specials = ["C", "+/-", "%"];
 
-export default { numbers, operators, specials };
+export { numbers, operators, specials };
